PRSD-704: Update integration tests
