+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/LandlordDeregistrationJourneyTests.kt
@@ -16,6 +16,7 @@ import org.springframework.test.context.jdbc.Sql
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.assertThat
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.LandlordDetailsPage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage.Companion.assertPageIs
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.landlordDeregistrationJourneyPages.ReasonFormPageLandlordDeregistration
 import uk.gov.communities.prsdb.webapp.services.LandlordService
 
 @Sql("/data-local.sql")
@@ -31,8 +32,7 @@ class LandlordDeregistrationJourneyTests : IntegrationTest() {
             .containsText("Are you sure you want to delete your account and all your properties on the database?")
         areYouSurePage.submitWantsToProceed()
 
-        // TODO PRSD-704 - redirect to reason page if the user with properties selects "yes"
-        assertPageIs(page, LandlordDetailsPage::class)
+        val reasonPage = assertPageIs(page, ReasonFormPageLandlordDeregistration::class)
     }
 
     @Test
